requirements/base.txt
python-slugify==8.0.1  # https://github.com/un33k/python-slugify
Pillow==9.5.0  # https://github.com/python-pillow/Pillow
rcssmin==1.1.1  # https://github.com/ndparker/rcssmin
argon2-cffi==21.3.0  # https://github.com/hynek/argon2_cffi
whitenoise==6.4.0  # https://github.com/evansd/whitenoise
redis==4.6.0  # https://github.com/redis/redis-py
hiredis==2.2.2  # https://github.com/redis/hiredis-py
# celery==5.2.7  # pyup: < 6.0  # https://github.com/celery/celery
celery==5.3.1  # pyup: < 6.0  # https://github.com/celery/celery
flower==2.0.0  # https://github.com/mher/flower

# Django
# ------------------------------------------------------------------------------
django==4.1.10
django-environ==0.10.0  # https://github.com/joke2k/django-environ
django-model-utils==4.3.1  # https://github.com/jazzband/django-model-utils
django-allauth==0.54.0  # https://github.com/pennersr/django-allauth
django-crispy-forms==2.0  # https://github.com/django-crispy-forms/django-crispy-forms
crispy-bootstrap5==0.7  # https://github.com/django-crispy-forms/crispy-bootstrap5
django-compressor==4.3.1  # https://github.com/django-compressor/django-compressor
django-redis==5.2.0  # https://github.com/jazzband/django-redis

# Django REST
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2  # https://django-rest-framework-simplejwt.readthedocs.io/en/latest/

# Django celery
# ------------------------------------------------------------------------------
django-celery-beat==2.5.0  # https://github.com/celery/django-celery-beat
django_celery_results==2.5.1


# Wagtail
# ------------------------------------------------------------------------------
wagtail==4.2.3  # https://github.com/wagtail/wagtail

# Wagtail Recaptcha
# ------------------------------------------------------------------------------
django-recaptcha==3.0.0
wagtail-django-recaptcha==1.0

# Wagtail Recaptcha
# ------------------------------------------------------------------------------
django-recaptcha==3.0.0
wagtail-django-recaptcha==1.0

# Wagtail Menu
# ------------------------------------------------------------------------------
wagtailmenus==3.1.8

# Wagtail Localize
# ------------------------------------------------------------------------------
wagtail-localize==1.5

# Wagtail-Autocomplete
# https://github.com/wagtail/wagtail-autocomplete
# ------------------------------------------------------------------------------
wagtail-autocomplete==0.10.0

# DSM Minio
# ------------------------------------------------------------------------------
minio==7.1.15

# Upload
# ------------------------------------------------------------------------------
lxml==4.9.3 # https://github.com/lxml/lxml
-e git+https://github.com/scieloorg/packtools#egg=packtools
-e git+https://github.com/scieloorg/scielo_scholarly_data#egg=scielo_scholarly_data

# DSM Publication
# ------------------------------------------------------------------------------
-e git+https://github.com/scieloorg/opac_schema.git@v2.66#egg=opac_schema
mongoengine==0.24.2
pymongo==3.12.0
aiohttp==3.8.5
python-magic==0.4.27

# DSM Migration
# ------------------------------------------------------------------------------
-e git+https://github.com/scieloorg/scielo_migration.git@1.1.0#egg=scielo_classic_website
python-dateutil==2.8.2
tornado>=6.3.2 # not directly required, pinned by Snyk to avoid a vulnerability

# Tenacity
# ------------------------------------------------------------------------------
tenacity==8.2.2  # https://pypi.org/project/tenacity/
urllib3==2.0.4
